Add CourseDatesDialog component with comprehensive date management features refactor: extract CourseDatesDialog from CourseList
